# api/openapi.yml
openapi: 3.0.0
info:
  description: |
    Payments is a simple API allowing consumers to manipulate collection of Payments.

    Since the structure of the Payment resource is domain specific and out of scope for this project, we chose to not validate the attributes part of the Payment object (see below) and handle it as an arbitrary JSON object. We validate the version field to be 0, which allows us to add any attributes validation in the future for next versions of the Payment resource. We also validate the id and organisation_id fields to be valid UUIDs.

    As for the errors, it may be necessary to introduce error codes in the future, but currently it is not needed as we can easily cover the errors via HTTP codes. However we always return a JSON representation of an error which can easily be extended by adding an error code.

  version: '0.0.1'
  title: Payments
  contact:
    email: 'dusan@kasan.sk'
  license:
    name: 'Apache 2.0'
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
  /payments:
    post:
      tags:
        - payments
      summary: Create a new Payment
      operationId: addPayment
      requestBody:
        description: Payment object that needs to be added to the Payments collection
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Payment'
      responses:
        201:
          description: Payment successfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        409:
          $ref: '#/components/responses/PaymentIdAlreadyExists'
        400:
          $ref: '#/components/responses/InvalidPayment'
        500:
          $ref: '#/components/responses/InternalServerError'
    get:
      tags:
        - payments
      summary: Fetch a window into the Payments collection sorted by id alphabetically. Returns a maximum of 100 Payments.
      operationId: getPayments
      parameters:
        - name: afterId
          in: query
          required: false
          description: If provided, the window will start with the first Payment with ID alphabetically higher than afterId.
          schema:
            type: string
            format: uuid
            example: '3fa85f64-5717-4562-b3fc-2c963f66afa6'
      responses:
        200:
          description: No error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentsCollection'
        500:
          $ref: '#/components/responses/InternalServerError'
  /payments/{id}:
    get:
      summary: Fetch a single payment by its unique identifier.
      operationId: getPayment
      tags:
        - payments
      parameters:
        - name: id
          in: path
          description: ID of payment to fetch
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Payment found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        404:
          $ref: '#/components/responses/PaymentNotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
    put:
      summary: Update a single payment by its unique identifier.
      description: This will overwrite all the payment data. A partial update mechanism (via PATCH) may be implemented later. The request data must not contain payment ID.
      operationId: updatePayment
      tags:
        - payments
      parameters:
        - name: id
          in: path
          description: ID of payment to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: PaymentWithoutId object that contains all the Payment data except id, to overwrite the Payment with the specified id
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentWithoutId'
      responses:
        200:
          description: Payment updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        404:
          $ref: '#/components/responses/PaymentNotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Remove a single payment by its unique identifier.
      operationId: deletePayment
      tags:
        - payments
      parameters:
        - name: id
          in: path
          description: ID of payment to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Payment deleted.
        404:
          $ref: '#/components/responses/PaymentNotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
components:
  schemas:
    PaymentWithoutId:
      type: object
      properties:
        type:
          type: string
          enum:
            - 'Payment'
        version:
          type: 'integer'
          format: 'int64'
          enum:
            - 0
        organisation_id:
          type: string
          format: uuid
        attributes:
          type: object
      example:
        type: Payment
        version: 0
        organisation_id: 743d5b63-8e6f-432e-a8fa-c5d8d2ee5fcb
        attributes:
          amount: '100.21'
          beneficiary_party:
            account_name: W Owens
            account_number: '31926819'
            account_number_code: BBAN
            account_type: 0
            address: 1 The Beneficiary Localtown SE2
            bank_id: '403000'
            bank_id_code: GBDSC
            name: Wilfred Jeremiah Owens
          charges_information:
            bearer_code: SHAR
            sender_charges:
              - amount: '5.00'
                currency: GBP
              - amount: '10.00'
                currency: USD
            receiver_charges_amount: '1.00'
            receiver_charges_currency: USD
          currency: GBP
          debtor_party:
            account_name: EJ Brown Black
            account_number: GB29XABC10161234567801
            account_number_code: IBAN
            address: 10 Debtor Crescent Sourcetown NE1
            bank_id: '203301'
            bank_id_code: GBDSC
            name: Emelia Jane Brown
          end_to_end_reference: Wil piano Jan
          fx:
            contract_reference: FX123
            exchange_rate: '2.00000'
            original_amount: '200.42'
            original_currency: USD
          numeric_reference: '1002001'
          payment_id: '123456789012345678'
          payment_purpose: Paying for goods/services
          payment_scheme: FPS
          payment_type: Credit
          processing_date: '2017-01-18'
          reference: Payment for Em's piano lessons
          scheme_payment_sub_type: InternetBanking
          scheme_payment_type: ImmediatePayment
          sponsor_party:
            account_number: '56781234'
            bank_id: '123123'
            bank_id_code: GBDSC
    Payment:
      allOf:
        - $ref: '#/components/schemas/PaymentWithoutId'
        - type: object
          properties:
            id:
              type: string
              format: uuid
          example:
            type: Payment
            id: 4ee3a8d8-ca7b-4290-a52c-dd5b6165ec43
            version: 0
            organisation_id: 743d5b63-8e6f-432e-a8fa-c5d8d2ee5fcb
            attributes:
              amount: '100.21'
              beneficiary_party:
                account_name: W Owens
                account_number: '31926819'
                account_number_code: BBAN
                account_type: 0
                address: 1 The Beneficiary Localtown SE2
                bank_id: '403000'
                bank_id_code: GBDSC
                name: Wilfred Jeremiah Owens
              charges_information:
                bearer_code: SHAR
                sender_charges:
                  - amount: '5.00'
                    currency: GBP
                  - amount: '10.00'
                    currency: USD
                receiver_charges_amount: '1.00'
                receiver_charges_currency: USD
              currency: GBP
              debtor_party:
                account_name: EJ Brown Black
                account_number: GB29XABC10161234567801
                account_number_code: IBAN
                address: 10 Debtor Crescent Sourcetown NE1
                bank_id: '203301'
                bank_id_code: GBDSC
                name: Emelia Jane Brown
              end_to_end_reference: Wil piano Jan
              fx:
                contract_reference: FX123
                exchange_rate: '2.00000'
                original_amount: '200.42'
                original_currency: USD
              numeric_reference: '1002001'
              payment_id: '123456789012345678'
              payment_purpose: Paying for goods/services
              payment_scheme: FPS
              payment_type: Credit
              processing_date: '2017-01-18'
              reference: Payment for Em's piano lessons
              scheme_payment_sub_type: InternetBanking
              scheme_payment_type: ImmediatePayment
              sponsor_party:
                account_number: '56781234'
                bank_id: '123123'
                bank_id_code: GBDSC
...
    Error:
      type: object
      properties:
        message:
          type: string
      required:
        - message
  responses:
    PaymentIdAlreadyExists:
      description: Payment with the specified ID already exists.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - properties:
                  message:
                    enum:
                      - payment with the specified ID already exists
    PaymentNotFound:
      description: Payment with the specified ID not found.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - properties:
                  message:
                    enum:
                      - payment with the specified ID not found
    InvalidPayment:
      description: Invalid Payment received.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            invalid version:
              value:
                message: invalid version number, must be 0
            invalid type:
              value:
                message: invalid type, must be 'Payment'
            invalid id:
              value:
                message: invalid id, must be a uuid
            invalid organisation_id:
              value:
                message: invalid organisation_id, must be a uuid
    InternalServerError:
      description: Unknown internal server error.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - properties:
                  message:
                    enum:
                      - internal server error